requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
openai>=1.3.0
httpx>=0.25.0
fastapi>=0.110.0
uvicorn>=0.23.0
music21>=8.3.0
abjad>=3.14.0
librosa>=0.10.0
numpy>=1.21.0
Pillow>=9.0.0
pandas>=1.3.0
pillow>=9.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.15.0
requests>=2.28.0
selenium>=4.15.0
beautifulsoup4>=4.12.0 